Enhance SNR calculations in common.py and simulate_inject.py: support multiple SNR definitions and improve trail injection logic

+++ ADCNN/data/dataset_creation/simulate_inject.py
@@ -56,7 +56,7 @@ def generate_one_line(n_inject, trail_length, mag, beta, ref, dimensions, seed, 
         names=('injection_id', 'ra', 'dec', 'source_type', 'trail_length', 'mag', 'beta', 'visit', 'detector',
                'integrated_mag', 'PSF_mag', 'SNR', 'physical_filter', 'x', 'y', 'stack_model_SNR'),
         dtype=('int64', 'float64', 'float64', 'str', 'float64', 'float64', 'float64', 'int64', 'int64', 'float64',
-               'float64', 'float64', 'str', 'int64', 'int64', 'int64'))
+               'float64', 'float64', 'str', 'int64', 'int64', 'float64'))
 
     H, W = int(dimensions.y), int(dimensions.x)
     if forbidden_mask is None:
